Refactor round 2: remove pygame dependancy, remove unused files
