+++ source/env_generator.py
@@ -1,110 +1,13 @@
 # python file to random generate environments
 from loguru import logger
-from env_config import EnvConfig
-from robot_base_env import RobotBaseEnv, pd_controller
-from robot_base_config import RobotBaseCBFConfig
-from robot_base import RobotBase
-from cbf_costmap import CBFCostmap
-from cbf_infused_a_star import CBFInfusedAStar
-from planner_comparison import PlannerComparison
-from a_star import AStarPlanner
-from visualization import VisualizeSimulation
-from cbfpy import CBF
 import random
 from obstacles import RectangleObstacle, CircleObstacle
 import numpy as np
 import os
 import json
-from perception import Perception, Sensor
-
-
-class EnvGeneratorConfig:
-    def __init__(
-        self,
-        number_of_simulations: int,
-        environment_size: tuple,
-        grid_size: float,
-        max_obstacle_size: dict,
-        max_duration_of_simulation: float,
-        min_goal_distance: float,
-        work_dir: str,
-        use_safety_margin: bool,
-        max_sensor_noise: float,
-        cbf_state_uncertainty_mode: str,
-        fps=60,
-        robot_size=(1, 1),
-        min_number_of_obstacles=1,
-        max_number_of_obstacles=1,
-        min_number_of_sensors=1,
-        max_number_of_sensors=1,
-        cbf_reduction="min",
-        cbf_infused_a_star=False,
-        planner_comparison=False,
-    ):
-        # general parameters
-        self.number_of_simulations = number_of_simulations
-        self.max_duration_of_simulation = max_duration_of_simulation  # in seconds
-        self.min_goal_distance = min_goal_distance  # in m
-        self.work_dir = work_dir
-        self.fps = fps
-        self.planner_comparison = planner_comparison  # bool
-
-        # environment parameters (robot and obstacles)
-        self.environment_size = environment_size  # in m
-        self.robot_size = robot_size  # in m
-        self.min_number_of_obstacles = min_number_of_obstacles
-        self.max_number_of_obstacles = max_number_of_obstacles
-        self.max_obstacle_size = max_obstacle_size  # in m
-
-        # costmap parameters
-        self.grid_size = grid_size  # in m
-
-        # cbf parameters
-        self.cbf_state_uncertainty_mode = (
-            cbf_state_uncertainty_mode  # choose between robust or probabilistic
-        )
-        self.use_safety_margin = use_safety_margin  # bool whether to use safety margin
-        self.cbf_reduction = cbf_reduction  # the reduction to get the cbf in one grid, options: ['min', 'mean', 'sum']
-        self.cbf_infused_a_star = cbf_infused_a_star  # bool
-
-        # perception parameters
-        self.min_number_of_sensors = min_number_of_sensors
-        self.max_number_of_sensors = max_number_of_sensors
-        self.max_sensor_noise = max_sensor_noise  # in m
-
-    def log_information(self):
-        logger.info(f"Workdir: {self.work_dir}")
-        logger.info(f"Environment size: {self.environment_size}")
-        logger.info(f"FPS: {self.fps}")
-        logger.info(f"Grid size: {self.grid_size} m")
-        logger.info(f"Max duration for simulation: {self.max_duration_of_simulation} s")
-        logger.info(f"Minimum distance to the goal: {self.min_goal_distance} m")
-        logger.info(f"Robot size: {self.robot_size}")
-        logger.info(f"Use safety margin: {self.use_safety_margin}")
-        logger.info(f"CBF reduction mode: {self.cbf_reduction}")
-        logger.info(f"CBF infused A*: {self.cbf_infused_a_star}")
-        logger.info(f"CBF state uncertainty mode: {self.cbf_state_uncertainty_mode}")
-        logger.info(f"Max  noise (standard deviation): {self.max_sensor_noise}")
-        logger.info(f"Min number of sensors: {self.min_number_of_sensors}")
-        logger.info(f"Max number of sensors: {self.max_number_of_sensors}")
-        logger.info(f"Min number of obstascles: {self.min_number_of_obstacles}")
-        logger.info(f"Max number of obstascles: {self.max_number_of_obstacles}")
-        for key, val in self.max_obstacle_size.items():
-            logger.info(f"Max obstacle size for {key}: {val} m")
-
-    def save_to_file(self, path: str):
-        # Convert all class attributes to a serializable dict
-        config_dict = self.__dict__.copy()
-        with open(path, "w") as f:
-            json.dump(config_dict, f, indent=4)
-        logger.success(f"Configuration saved to {path}")
-
-    @classmethod
-    def from_file(cls, path: str):
-        with open(path, "r") as f:
-            config_dict = json.load(f)
-        logger.success(f"Configuration loade from {path}")
-        return cls(**config_dict)
+from perception import Sensor
+from robot import Robot
+from env_generator_config import EnvGeneratorConfig
 
 
 class EnvGenerator:
@@ -112,21 +15,16 @@ class EnvGenerator:
     # and visualization for that
     def __init__(self, config: EnvGeneratorConfig):
         self.config = config
-        self.x_range = self.config.environment_size[0] / 2  # + - range for x dimenstion
-        self.y_range = self.config.environment_size[1] / 2  # + - range for y dimension
+        self.x_range = self.config.costmap_size[0] / 2  # + - range for x dimenstion
+        self.y_range = self.config.costmap_size[1] / 2  # + - range for y dimension
 
         # create the env_config (mainly needed for switching the x and y dimension)
-        self.env_config = EnvConfig(
-            pixels_per_meter=50 * np.array([1, -1]), screen_width=800, screen_height=800
-        )
-        self.pygame_screen = False  # we dont want the pygame screen to pops up
         self.work_dir_available = self._create_work_dir()
 
-        # create workdir to save all the combined image (only if there is no comparison)
-        if not self.config.planner_comparison:
-            os.makedirs(
-                f"{self.config.work_dir}/simulation_results/all_envs", exist_ok=True
-            )
+        # create workdir to save all the combined image
+        os.makedirs(
+            f"{self.config.work_dir}/simulation_results/all_envs", exist_ok=True
+        )
 
         # log all the important information
         self.config.log_information()
@@ -155,14 +53,12 @@ class EnvGenerator:
 
             return False
 
-    def save_env_information(
-        self, dir, robot: RobotBase, obstacles: list, sensors: list
-    ):
+    def save_env_information(self, dir, robot: Robot, obstacles: list, sensors: list):
         # Prepare dictionary with only JSON-serializable data
         env_dict = {
-            "start_pos": robot.pos_center_start.tolist(),
-            "start_vel": robot.vel_center_start.tolist(),
-            "goal_pos": robot.pos_goal.tolist(),
+            "start_pos": robot.true_state[:2].tolist(),
+            "start_vel": robot.true_state[2:].tolist(),
+            "goal_pos": robot.goal_position.tolist(),
             "obstacles": [],
             "sensors": [],
         }
@@ -174,6 +70,7 @@ class EnvGenerator:
                         "type": "circle",
                         "center": obstacle.pos_center.tolist(),
                         "radius": obstacle.radius,
+                        "robot_radius": obstacle.robot_radius,
                     }
                 )
             elif isinstance(obstacle, RectangleObstacle):
@@ -183,6 +80,7 @@ class EnvGenerator:
                         "center": obstacle.pos_center.tolist(),
                         "height": obstacle.height,
                         "width": obstacle.width,
+                        "robot_radius": obstacle.robot_radius,
                     }
                 )
 
@@ -218,8 +116,7 @@ class EnvGenerator:
                     CircleObstacle(
                         pos_center=np.array(obs["center"]),
                         radius=obs["radius"],
-                        env_config=self.env_config,
-                        robot=None,
+                        robot_radius=obs["robot_radius"],
                         id=i,
                     )
                 )
@@ -229,8 +126,7 @@ class EnvGenerator:
                         pos_center=np.array(obs["center"]),
                         height=obs["height"],
                         width=obs["width"],
-                        env_config=self.env_config,
-                        robot=None,
+                        robot_radius=obs["robot_radius"],
                         id=i,
                     )
                 )
@@ -268,16 +164,31 @@ class EnvGenerator:
         )
         return max_goal
 
-    def _check_collision(self, robot, obstacles, generation=False):
+    def _check_collision(
+        self,
+        robot_pos: np.ndarray,
+        goal_pos: np.ndarray,
+        robot_width: float,
+        robot_height: float,
+        obstacles: list,
+        generation=False,
+    ):
         # checks if there is a collision between robot and obstacles
         obs_to_keep = []
         for i, obstacle in enumerate(obstacles):
             if generation:
                 # only add small safety margin when generating the obstacles -> prevent robot to be to close to obstacle
-                collision = obstacle.check_collision(robot, safety_margin=0.2)
+                collision = obstacle.check_collision(
+                    robot_pos, robot_width, robot_height, safety_margin=0.2
+                )
+                # only check for goal location when generating
+                goal_feasible = obstacle.check_goal_position(
+                    goal_pos, robot_width, robot_height
+                )
             else:
-                collision = obstacle.check_collision(robot)
-            goal_feasible = obstacle.check_goal_position(robot)
+                collision = obstacle.check_collision(
+                    robot_pos, robot_width, robot_height
+                )
 
             if generation:
                 if collision:
@@ -294,7 +205,7 @@ class EnvGenerator:
                     obs_to_keep.append(True)
             elif not generation and collision:
                 logger.error(
-                    f"Collision between robot and obstacle (id={obstacle.id})! Robot position: {robot.position}"
+                    f"Collision between robot and obstacle (id={obstacle.id})! Robot position: {robot_pos}"
                 )
 
         if generation:
@@ -303,8 +214,16 @@ class EnvGenerator:
             ]
             return filtered_obstacles
 
-    def _generate_obstacles(self, robot: RobotBase, max_tries=1000):
+    def _generate_obstacles(
+        self,
+        start_pos: np.ndarray,
+        goal_pos: np.ndarray,
+        robot_width: float,
+        robot_height: float,
+        max_tries=1000,
+    ):
         obstacles = []
+        robot_radius = np.linalg.norm(robot_width / 2 + robot_height / 2)
         for i in range(max_tries):
             number_of_obstacles = random.randint(
                 self.config.min_number_of_obstacles, self.config.max_number_of_obstacles
@@ -322,9 +241,9 @@ class EnvGenerator:
                     )
                     obstacle = CircleObstacle(
                         radius=radius,
+                        robot_radius=robot_radius,
                         pos_center=pos_center,
-                        env_config=self.env_config,
-                        robot=robot,
+                        robot=None,
                         id=i,
                     )
 
@@ -347,9 +266,9 @@ class EnvGenerator:
                     obstacle = RectangleObstacle(
                         width=width,
                         height=height,
+                        robot_radius=robot_radius,
                         pos_center=pos_center,
-                        env_config=self.env_config,
-                        robot=robot,
+                        robot=None,
                         id=i,
                     )
                     logger.info(
@@ -359,7 +278,14 @@ class EnvGenerator:
                 obstacles.append(obstacle)
 
             # check if they are in collision
-            obstacles = self._check_collision(robot, obstacles, generation=True)
+            obstacles = self._check_collision(
+                robot_pos=start_pos,
+                goal_pos=goal_pos,
+                robot_width=robot_width,
+                robot_height=robot_height,
+                obstacles=obstacles,
+                generation=True,
+            )
 
             if len(obstacles) > 0:
                 return obstacles
@@ -374,316 +300,112 @@ class EnvGenerator:
             # generate the robot object
             robot_x = np.round(random.uniform(-self.x_range, self.x_range), 2)
             robot_y = np.round(random.uniform(-self.y_range, self.y_range), 2)
-            pos_goal = self._generate_goal((robot_x, robot_y))
-
-            robot = RobotBase(
-                width=self.config.robot_size[0],
-                height=self.config.robot_size[1],
-                env_config=self.env_config,
-                pos_goal=pos_goal,
-                pos_center_start=np.array([robot_x, robot_y]),
-            )
-            logger.info(f"Robot start: {robot.position}, goal: {robot.pos_goal}")
+            start_pos = np.array([robot_x, robot_y])
+            goal_pos = self._generate_goal(start_pos)
 
             # generate the obstacles
-            obstacles = self._generate_obstacles(robot)
+            obstacles = self._generate_obstacles(
+                start_pos=start_pos,
+                goal_pos=goal_pos,
+                robot_width=self.config.robot_width,
+                robot_height=self.config.robot_height,
+            )
 
             # set sensors to None to create sensors
-            sensors = None
+            num_sensors = random.randint(
+                self.config.min_number_of_sensors, self.config.max_number_of_sensors
+            )
+            sensors = []
+            for _ in range(num_sensors):
+                sx = np.round(
+                    random.uniform(
+                        -self.config.costmap_size[0] / 2,
+                        self.config.costmap_size[0] / 2,
+                    ),
+                    2,
+                )
+                sy = np.round(
+                    random.uniform(
+                        -self.config.costmap_size[1] / 2,
+                        self.config.costmap_size[1] / 2,
+                    ),
+                    2,
+                )
+                sensors.append(Sensor(np.array([sx, sy])))
         else:
-            start_pos, start_vel, pos_goal, obstacles, sensors = (
+            start_pos, start_vel, goal_pos, obstacles, sensors = (
                 self.load_env_information(loaded_env_dir)
             )
-            robot = RobotBase(
-                width=self.config.robot_size[0],
-                height=self.config.robot_size[1],
-                env_config=self.env_config,
-                pos_goal=pos_goal,
-                pos_center_start=start_pos,
-                vel_center_start=start_vel,
-            )
 
-            # add robot to obstacles
-            for obstacle in obstacles:
-                obstacle.robot = robot
-
-        # create config and cbf
-        config = RobotBaseCBFConfig(obstacles)
-        cbf = CBF.from_config(config)
-
-        cbf_costmap = CBFCostmap(
-            costmap_size=self.config.environment_size,
+        # create the robot
+        initial_state = np.array([start_pos, start_vel]).flatten()
+        robot = Robot(
+            costmap_size=self.config.costmap_size,
             grid_size=self.config.grid_size,
-            cbf_config=config,
-            cbf_reduction=self.config.cbf_reduction,
-        )
-
-        # create perception module (costmap included)
-        num_sensors = random.randint(
-            self.config.min_number_of_sensors, self.config.max_number_of_sensors
-        )
-        perception = Perception(
-            costmap_size=self.config.environment_size,
-            grid_size=self.config.grid_size,
-            cbf=cbf,
-            num_sensors=num_sensors,
-            min_values_state=np.array([-10, -10, -1.5, -1.5]),
-            max_values_state=np.array([10, 10, 1.5, 1.5]),
+            planner_mode=self.config.planner_mode,
+            width=self.config.robot_width,
+            height=self.config.robot_height,
+            min_values_state=self.config.min_values_state,
+            max_values_state=self.config.max_values_state,
             max_sensor_noise=self.config.max_sensor_noise,
-            num_samples_per_dim=4,
+            cbf_state_uncertainty_mode=self.config.cbf_state_uncertainty_mode,
+            control_fps=self.config.control_fps,
+            state_estimation_fps=self.config.state_estimation_fps,
+            goal_tolerance=self.config.goal_tolerance,
+            Kp=self.config.Kp,
+            Kd=self.config.Kd,
+            u_min_max=self.config.u_min_max,
+            initial_state=initial_state,
             sensors=sensors,
-        )
-
-        # generate the planner
-        planners = {}
-        if self.config.cbf_infused_a_star:
-            planner = CBFInfusedAStar(
-                costmap_size=self.config.environment_size,
-                grid_size=self.config.grid_size,
-                obstacles=obstacles,
-                cbf_costmap=cbf_costmap,
-            )
-            planners["CBF infused A*"] = planner
-        else:
-            planner = AStarPlanner(
-                costmap_size=self.config.environment_size,
-                grid_size=self.config.grid_size,
-                obstacles=obstacles,
-            )
-            planners["A*"] = planner
-
-        if self.config.planner_comparison:
-            if isinstance(planner, CBFInfusedAStar):
-                other_planner = AStarPlanner(
-                    costmap_size=self.config.environment_size,
-                    grid_size=self.config.grid_size,
-                    obstacles=obstacles,
-                )
-                planners["A*"] = other_planner
-            else:
-                other_planner = CBFInfusedAStar(
-                    costmap_size=self.config.environment_size,
-                    grid_size=self.config.grid_size,
-                    obstacles=obstacles,
-                    cbf_costmap=cbf_costmap,
-                )
-                planners["CBF infused A*"] = other_planner
-
-        # create environment
-        env = RobotBaseEnv(
-            env_config=self.env_config,
-            robot_base=robot,
             obstacles=obstacles,
-            pygame_screen=self.pygame_screen,
-            fps=self.config.fps,
         )
 
-        visualizers = {}
-        for key in planners.keys():
-            visualizer = VisualizeSimulation(
-                pos_goal=robot.pos_goal, obstacles=obstacles, show_plot=False
-            )
-            visualizers[key] = visualizer
+        # add robot to obstacles
+        for obstacle in obstacles:
+            obstacle.robot = robot
 
-        return env, visualizers, config, cbf, planners, cbf_costmap, perception
+        # plan the path
+        robot.plan(goal_pos=goal_pos)
 
-    def _run_planner(
-        self,
-        env: RobotBaseEnv,
-        visualizer: VisualizeSimulation,
-        config: RobotBaseCBFConfig,
-        cbf: CBF,
-        planner: AStarPlanner | CBFInfusedAStar,
-        cbf_costmap: CBFCostmap,
-        perception: Perception,
-        env_folder: str,
-    ):
-        max_timesteps = env.fps * self.config.max_duration_of_simulation
-        timestep = 0
-
-        # add path and costmaps to visualizer
-        visualizer.data.path = env.robot_base.path
-        visualizer.data.planner_costmap = planner.compute_distance_map(
-            start=env.robot_base.position
-        )
-        visualizer.data.cbf_costmap = cbf_costmap.costmap
-        visualizer.data.perception_magnitude_costmap = (
-            perception.perception_magnitude_costmap
-        )
-        visualizer.data.noise_costmap = perception.noise_costmap
-        visualizer.data.sensor_positions = perception.sensor_positions
-
-        # run env
-        while env.running and timestep < max_timesteps:
-            current_state = env.get_state()
-            estimated_state = perception.get_estimated_state(current_state)
-
-            # calculate nominal control
-            # nominal_control = pd_controller(pos_des, estimated_state[:2], estimated_state[2:])
-            nominal_control = env.robot_base.velocity_pd_controller()
-            if np.isnan(current_state).any():
-                logger.warning(
-                    f"[{timestep}]: Nan value in current_state: {current_state}"
-                )
-                break
-
-            # calculate the safety margin
-            if self.config.use_safety_margin:
-                noise = perception.get_perception_noise(x_true=current_state[:2])
-                safety_margin = perception.calculate_safety_margin(
-                    noise=noise,
-                    u_nominal=nominal_control,
-                    mode=self.config.cbf_state_uncertainty_mode,
-                )
-            else:
-                safety_margin = np.zeros(config.num_obstacles)
-
-            # safe data for visualizer
-            # for h take the true state to calculate h
-            visualizer.data.timestep.append(timestep)
-            h_true = config.h_1(
-                current_state, np.zeros(config.num_obstacles)
-            )  # don't add safety margin here (because state is true)
-            h_true = config.alpha(h_true)
-            visualizer.data.h_true.append(np.array(h_true))
-            h_estimated = config.h_1(estimated_state, safety_margin)
-            h_estimated = config.alpha(h_estimated)
-            visualizer.data.h_estimated.append(np.array(h_estimated))
-            visualizer.data.robot_pos.append(current_state[:2])
-            visualizer.data.robot_pos_estimated.append(estimated_state[:2])
-            visualizer.data.robot_vel.append(current_state[2:])
-
-            # apply safety filter
-            u = cbf.safety_filter(estimated_state, nominal_control, safety_margin)
-
-            # safe control data for visualizer
-            visualizer.data.u_cbf.append(u)
-            visualizer.data.u_nominal.append(nominal_control)
-            visualizer.data.safety_margin.append(safety_margin)
-
-            # change environment
-            env.apply_control(u)
-            self._check_collision(
-                robot=env.robot_base, obstacles=env.obstacles, generation=False
-            )
-            env.step()
-
-            # increment timestep
-            timestep += 1
-
-        # check the maximum tolerance as an effect of the state estimation uncertainty
-        if env.robot_base.check_goal_reached(
-            tolerance=self.config.grid_size + self.config.max_sensor_noise + 0.01
-        ):
-            logger.success(f"Goal reached in {timestep} timesteps.")
-            goal_reached = True
-        else:
-            logger.warning(
-                f"Simulation ended without reaching the goal. Timesteps: {timestep}"
-            )
-            goal_reached = False
-
-        # add last information for visualizer
-        visualizer.data.timestep.append(timestep)
-        visualizer.data.u_cbf.append(u)
-        visualizer.data.u_nominal.append(nominal_control)
-        visualizer.data.safety_margin.append(safety_margin)
-        h_true = config.h_1(current_state, np.zeros(config.num_obstacles))
-        h_true = config.alpha(h_true)
-        visualizer.data.h_true.append(np.array(h_true))
-        h_estimated = config.h_1(estimated_state, safety_margin)
-        h_estimated = config.alpha(h_estimated)
-        visualizer.data.h_estimated.append(np.array(h_estimated))
-        visualizer.data.robot_pos.append(current_state[:2])
-        visualizer.data.robot_pos_estimated.append(estimated_state[:2])
-        visualizer.data.robot_vel.append(current_state[2:])
-
-        # set planner filenames
-        if isinstance(planner, CBFInfusedAStar):
-            planner_filename = "cbf_infused_a_star"
-        else:
-            planner_filename = "a_star"
-
-        # save the data
-        visualizer.data.save_data(dir=f"{env_folder}/{planner_filename}")
-
-        # generate drawings
-        if goal_reached:
-            filename = f"{env_folder}/{planner_filename}_success.png"
-        else:
-            filename = f"{env_folder}/{planner_filename}_fail.png"
-
-        filename = [filename]
-        if not self.config.planner_comparison:
-            env_number = env_folder.split("_")[-1]
-            if goal_reached:
-                _filename = f"{self.config.work_dir}/simulation_results/all_envs/{env_number}_success.png"
-            else:
-                _filename = f"{self.config.work_dir}/simulation_results/all_envs/{env_number}_fail.png"
-
-            filename.append(_filename)
-
-        visualizer.create_full_plot(planner, filename)
-
-        # return whether the goal is reached
-        return goal_reached
+        return robot, obstacles, sensors
 
     @logger.catch
     def _run_env(self, env_folder, loaded_env_dir=None):
-        env, visualizers, config, cbf, planners, cbf_costmap, perception = (
-            self._generate_env_elements(loaded_env_dir)
-        )
-        # perception.add_sensor(Sensor(sensor_position=np.array([6, -1])))
-        success = {}
+        robot, obstacles, sensors = self._generate_env_elements(loaded_env_dir)
+
+        # here you can add sensors to the perception module of the robot
+        # robot.perception.add_sensor(Sensor(sensor_position=np.array([4, 1])))
 
         # save the environment parameters
         self.save_env_information(
             dir=env_folder,
-            robot=env.robot_base,
-            obstacles=env.obstacles,
-            sensors=perception.sensors,
+            robot=robot,
+            obstacles=obstacles,
+            sensors=sensors,
         )
 
-        for planner_name, planner in planners.items():
-            logger.info(f"Planner: {planner_name}")
+        # run the simulation
+        sim_output = robot.run_simulation(
+            sim_time=self.config.max_duration_of_simulation,
+            env_folder=env_folder
+        )
 
-            # plan the path
-            env.robot_base.reset()
-            path = planner.plan(
-                start_coords=env.robot_base.position,
-                goal_coords=env.robot_base.pos_goal,
-            )
+        # create plot
+        env_number = env_folder.split("_")[-1]
+        # generate drawings
+        if sim_output:
+            filenames = [
+                f"{env_folder}/simulation_success.png",
+                f"{self.config.work_dir}/simulation_results/all_envs/{env_number}_success.png",
+            ]
+        else:
+            filenames = [
+                f"{env_folder}/simulation_fail.png",
+                f"{self.config.work_dir}/simulation_results/all_envs/{env_number}_fail.png",
+            ]
 
-            # if no path go to the next planner
-            if path is None:
-                logger.error(f"No path found! Skip this planner!")
-                success[planner_name] = None
-                continue
-
-            # path found -> add the path, activate the environment and run the planner
-            env.robot_base.add_path(path=planner.path_world)
-            env.running = True
-            goal_reached = self._run_planner(
-                env=env,
-                visualizer=visualizers[planner_name],
-                config=config,
-                cbf=cbf,
-                planner=planner,
-                cbf_costmap=cbf_costmap,
-                perception=perception,
-                env_folder=env_folder,
-            )
-            success[planner_name] = goal_reached
-
-        if self.config.planner_comparison:
-            planner_comparison = PlannerComparison(
-                a_star_planner=planners["A*"],
-                cbf_a_star_planner=planners["CBF infused A*"],
-                visualizers=visualizers,
-            )
-            planner_comparison.plot_comparison(f"{env_folder}/planner_comparison.png")
-
-        return success
+        robot.plot(filenames)
+        return sim_output
 
     def run_env_from_file(self, env_file: str, env_folder: str):
         # run an environment from a file
@@ -743,31 +465,35 @@ class EnvGenerator:
 
 
 def main():
-    directory = "./runs/debug_1"
+    directory = "./runs/debug_refactor"
 
     # config = EnvGeneratorConfig(
     #     number_of_simulations=1,
-    #     fps=50,
-    #     min_number_of_obstacles=5,
-    #     max_number_of_obstacles=10,
-    #     environment_size=(20, 20),
-    #     grid_size=0.1,
-    #     max_obstacle_size={
-    #         'circle': 3.0,
-    #         'rectangle': (3.0, 3.0)
-    #     },
+    #     work_dir=directory,
     #     max_duration_of_simulation=20,
     #     min_goal_distance=15,
-    #     robot_size=(1, 1),
-    #     use_safety_margin=True,
-    #     cbf_reduction='min',
-    #     work_dir=directory,
-    #     cbf_infused_a_star=True,
-    #     planner_comparison=False,
+    #     min_number_of_obstacles=5,
+    #     max_number_of_obstacles=10,
+    #     max_obstacle_size={"circle": 3.0, "rectangle": [3.0, 3.0]},
+    #     min_number_of_sensors=1,
     #     max_number_of_sensors=5,
-    #     max_sensor_noise=0.1
+    #     costmap_size=np.array([20, 20]),
+    #     grid_size=0.1,
+    #     planner_mode="CBF infused A*",
+    #     robot_width=1.0,
+    #     robot_height=1.0,
+    #     min_values_state=np.array([-10, -10, -1.5, -1.5]),
+    #     max_values_state=np.array([10, 10, 1.5, 1.5]),
+    #     max_sensor_noise=0.1,
+    #     cbf_state_uncertainty_mode="robust",
+    #     control_fps=50,
+    #     state_estimation_fps=50,
+    #     goal_tolerance=0.1,
+    #     Kp=0.5,
+    #     Kd=0.1,
+    #     u_min_max=np.array([-1000, 1000])
     # )
-    config = EnvGeneratorConfig.from_file("./runs/debug/env_config.json")
+    config = EnvGeneratorConfig.from_file("./runs/baseline_hard/env_config.json")
 
     # create logger
     logger.add(f"{config.work_dir}/simulations.log", rotation="10 MB")
@@ -777,7 +503,7 @@ def main():
 
     # apply same environment for debugging
     envs.run_env_from_file(
-        env_file="./runs/debug/simulation_results/env_6/env_data.json",
+        env_file="./runs/baseline_hard/simulation_results/loaded_env/env_data.json",
         env_folder="loaded_env",
     )
